src/test/java/uk/gov/hmcts/opal/service/legacy/LegacyLogAuditDetailServiceTest.java
package uk.gov.hmcts.opal.service.legacy;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.slf4j.Logger;
import uk.gov.hmcts.opal.dto.search.LogAuditDetailSearchDto;
import uk.gov.hmcts.opal.entity.LogAuditDetailEntity;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertThrows;

@ExtendWith(MockitoExtension.class)
class LegacyLogAuditDetailServiceTest extends LegacyTestsBase {

    @Mock
    private Logger log;

    @InjectMocks
    private LegacyLogAuditDetailService legacyLogAuditDetailService;

    @Test
    void testGetLogAuditDetail() {
        // Arrange

        LogAuditDetailEntity logAuditDetailEntity = LogAuditDetailEntity.builder().build();

        // Act
        LegacyGatewayResponseException exception = assertThrows(
            LegacyGatewayResponseException.class,
            () -> legacyLogAuditDetailService.getLogAuditDetail(1)
        );

        // Assert
        assertNotNull(legacyLogAuditDetailService.getLog());
        assertNotNull(exception);
        assertEquals(NOT_YET_IMPLEMENTED, exception.getMessage());

    }

    @Test
    void testSearchLogAuditDetails() {
        // Arrange
        LogAuditDetailSearchDto searchDto = LogAuditDetailSearchDto.builder().build();

        // Act
        LegacyGatewayResponseException exception = assertThrows(
            LegacyGatewayResponseException.class,
            () -> legacyLogAuditDetailService.searchLogAuditDetails(searchDto)
        );

        // Assert
        assertNotNull(exception);
        assertEquals(NOT_YET_IMPLEMENTED, exception.getMessage());

    }

}
